refactor: deleted _playlist and playlist variables, init block, and refreshDataFromNetwork() function
feat: added a database variable and assigned it to the getDatabase(), by passing application. Also defined videosRepository by calling the constructor and passing in the database. Also added an init block and launched a coroutine to call videosRepository.refreshVideos(). Also got the videos from the repository and assigned it to a playlist variable.